+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.1
+
+* Fix: [iOS] Race condition ending call remotely and locally results in a "End Call Failed: The operation couldn’t be completed." error. [Issue #184](https://github.com/cybex-dev/twilio_voice/issues/184)
+* Fix: [iOS] Plugin messages sent via Event Sink throwing an error due to sending from background thread. [Issue #256](https://github.com/cybex-dev/twilio_voice/issues/256)
+
 ## 0.2.0+1
 
 * Fix: [Web] `window.localStorage` deprecated members